requirements.txt
flask>=3.0.0
feedparser>=6.0.0
requests>=2.31.0
stripe>=7.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
gunicorn>=21.2.0
